Code cleanup, format and rename variables to match camel case. Convert loops into streams and data classes to records with updated jdk 17

+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/wired/addons/WiredAddonColourWheel.java
@@ -9,7 +9,7 @@ import com.cometproject.api.game.utilities.RandomUtil;
 
 
 public class WiredAddonColourWheel extends RoomItemFloor {
-    private static final int TIMEOUT = 4;
+    private static final double TIMEOUT = 4;
 
     public WiredAddonColourWheel(RoomItemData itemData, Room room) {
         super(itemData, room);
